CollectionBuilder renamed to match refactor
